vo: fix forgotten debug code
This was not intended to be committed in 26151062d86c8a86d1. It disables
the extra wakeup if working==true. I've convinced myself that the wakeup
was really needed at the time, so no idea how I didn't notice this until
someone pointed it out on the commit diff on github (lol).